backend/src/inheir_backend/helpers/auth.py
import logging
from ..config import AppConfig
import time
from typing import Dict
import jwt
import bcrypt
from pydantic import BaseModel

config: AppConfig = AppConfig()


class JwtPayload(BaseModel):
    user_id: str
    username: str
    expires: int
    role: str = "User"


def get_hashed_password(password: str) -> str:
    return bcrypt.hashpw(password.encode("utf-8"), bcrypt.gensalt()).decode("utf-8")


def verify_password(password: str, hashed_pass: str) -> bool:
    return bcrypt.checkpw(
        password.encode("utf-8"),
        hashed_pass.encode("utf-8"),
    )

def sign_jwt(user_id: str, username: str, role: str):
    try:
        payload = JwtPayload(
            user_id=user_id, username=username, expires=int(time.time() + 3600), role=role
        )
        token = jwt.encode(
            payload.dict(), config.env.jwt_secret, algorithm="HS512")
        return (token, payload.expires)
    except Exception as e:
        logging.error(e)



def decode_jwt(token: str) -> dict:
    try:
        decoded_token = jwt.decode(
            token, config.env.jwt_secret, algorithms=["HS512"])
        return (
            decoded_token if int(decoded_token["expires"]) >= int(
                time.time()) else None
        )
    except Exception as e:
        return None


def verify_jwt(token: str) -> bool:
    if decode_jwt(token):
        return True
    return False