backend/pyproject.toml
[project]
name = "inheir-backend"
version = "0.1.0"
description = "Backend for InHeir.AI"
authors = [{ name = "InHeir.AI" }]
requires-python = ">=3.13"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "fastapi[standard]>=0.115.6,<0.116",
    "motor>=3.6.1,<4",
    "pyjwt>=2.10.1,<3",
    "pydantic[email]>=2.10.5,<3",
    "pydantic-settings>=2.7.1,<3",
    "python-dotenv>=1.0.1,<2",
    "uvicorn>=0.34.0,<0.35",
    "azure-identity>=1.19.0,<2",
    "langchain>=0.3.20,<0.4",
    "openai[azure]>=1.66.3,<2",
    "azure-core>=1.32.0,<2",
    "azure-ai-generative>=1.0.0b11,<2",
    "langchain-openai>=0.3.8,<0.4",
    "langchain-community>=0.3.19,<0.4",
    "azure-ai-evaluation==1.1.0",
    "azure-ai-textanalytics>=5.3.0,<6",
    "marshmallow<3.20",
    "azure-cognitiveservices-speech>=1.43.0,<2",
    "aiofiles>=24.1.0,<25",
    "azure-storage-blob>=12.25.0,<13",
    "azure-ai-formrecognizer>=3.3.3,<4",
    "python-multipart>=0.0.20,<0.0.21",
    "ffmpeg-python>=0.2.0,<0.3",
    "azure-search-documents>=11.5.2,<12",
    "azure-search>=1.0.0b2,<2",
    "geopandas>=1.1.0,<2",
    "geopy>=2.4.1,<3",
    "tenacity>=9.1.2,<10",
    "bcrypt>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
